% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Catalogs.R
\name{get_Catalogs}
\alias{get_Catalogs}
\title{get_catalogs}
\usage{
get_Catalogs(x)
}
\arguments{
\item{x}{a comma separated list of catalogs you want downloaded}
}
\value{
list of
}
\description{
This function will return \code{\link{data.table}} containing the catalog or list of catalogs
             supplied
}
\examples{
\dontrun{
get_Catalogs("COUNTRY.EUSRrepCountry")
}
}
